Deprecate use of reserved word `Numeric`
This may fail with newer PHP versions.
