+++ bootstrap.php
@@ -8,4 +8,4 @@ if ( defined( 'MWSTAKE_MEDIAWIKI_COMPONENT_DATASTORE_VERSION' ) ) {
 	return;
 }
 
-define( 'MWSTAKE_MEDIAWIKI_COMPONENT_DATASTORE_VERSION', '1.0.2' );
+define( 'MWSTAKE_MEDIAWIKI_COMPONENT_DATASTORE_VERSION', '1.0.3' );
